Refactors PathBuilder and xhr method now just takes it as an argument so that custom PathBuilder impl instances can be used case by case
