+++ src/xhr/src/PathBuilderExample.ts
@@ -1,0 +1,13 @@
+import PathBuilder from './PathBuilder'
+
+export default class PathBuilderExample extends PathBuilder {
+  public posts() {
+    this.pathSegments.push('posts')
+    return this
+  }
+  public delete(id) {
+    this.pathSegments.push('posts')
+    this.id(id)
+    return this
+  }
+}
